Fixed a few bugs, and added a few new test roms.
The drawing (0xDXYN) did sometimes get out of range of the screen array, so if that happens, just ignore the new data and move on. The reg_load (0xFX65) did not load the registers correctly, but now it does. Increased the size of the memory data array. Catch FileNotFoundError instead of crashing. Changed the way the python modules are loaded for better readability.